Change package back to com.adaptris.aws.kinesis Bump docker-compose.yaml to 2.3.2
