--- interlok-aws-kinesis-sdk/src/main/java/com/adaptris/aws2/kinesis/RequestBuilder.java
@@ -1,10 +1,0 @@
-package com.adaptris.aws2.kinesis;
-
-import com.adaptris.core.AdaptrisMessage;
-import com.adaptris.core.CoreException;
-import com.amazonaws.services.kinesis.model.PutRecordsRequestEntry;
-
-public interface RequestBuilder {
-
-  Iterable<PutRecordsRequestEntry> build(String partitionKey, AdaptrisMessage message) throws CoreException;
-}
