+++ interlok-aws2-kinesis-sdk/docker-compose.yaml
@@ -1,4 +1,4 @@
-version: "2.2"
+version: "2.3.2"
 services:
   localstack:
     image: localstack/localstack
