+++ interlok-aws2-kinesis-sdk/src/main/java/com/adaptris/aws/kinesis/DefaultRequestBuilder.java
@@ -1,10 +1,11 @@
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import java.nio.ByteBuffer;
 import java.util.Collections;
 import com.adaptris.core.AdaptrisMessage;
-import com.amazonaws.services.kinesis.model.PutRecordsRequestEntry;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
+import software.amazon.awssdk.core.SdkBytes;
+import software.amazon.awssdk.services.kinesis.model.PutRecordsRequestEntry;
 
 /**
  * The default request builder.
@@ -20,9 +21,10 @@ public class DefaultRequestBuilder implements RequestBuilder {
 
   @Override
   public Iterable<PutRecordsRequestEntry> build(String partitionKey, AdaptrisMessage message) {
-    PutRecordsRequestEntry putRecordsRequestEntry  = new PutRecordsRequestEntry();
-    putRecordsRequestEntry.setData(ByteBuffer.wrap(message.getPayload()));
-    putRecordsRequestEntry.setPartitionKey(partitionKey);
+    PutRecordsRequestEntry putRecordsRequestEntry  = PutRecordsRequestEntry.builder()
+      .data(SdkBytes.fromByteBuffer(ByteBuffer.wrap(message.getPayload())))
+      .partitionKey(partitionKey)
+      .build();
     return Collections.singletonList(putRecordsRequestEntry);
   }
 }
