+++ interlok-aws2-kinesis-sdk/src/main/java/com/adaptris/aws/kinesis/KinesisSDKStreamProducer.java
@@ -1,4 +1,4 @@
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,14 +16,14 @@ import com.adaptris.core.util.Args;
 import com.adaptris.core.util.ExceptionHelper;
 import com.adaptris.interlok.util.CloseableIterable;
 import com.adaptris.util.NumberUtils;
-import com.amazonaws.services.kinesis.AmazonKinesis;
-import com.amazonaws.services.kinesis.model.PutRecordsRequest;
-import com.amazonaws.services.kinesis.model.PutRecordsRequestEntry;
-import com.amazonaws.services.kinesis.model.PutRecordsResult;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import software.amazon.awssdk.services.kinesis.KinesisClient;
+import software.amazon.awssdk.services.kinesis.model.PutRecordsRequest;
+import software.amazon.awssdk.services.kinesis.model.PutRecordsRequestEntry;
+import software.amazon.awssdk.services.kinesis.model.PutRecordsResponse;
 
 /**
  * Producer to amazon kinesis using the SDK.
@@ -113,7 +113,7 @@ public class KinesisSDKStreamProducer extends ProduceOnlyProducerImp {
   @Override
   protected void doProduce(AdaptrisMessage msg, String endpoint) throws ProduceException {
     try {
-      AmazonKinesis kinesisClient = retrieveConnection(AWSKinesisSDKConnection.class).kinesisClient();
+      KinesisClient kinesisClient = retrieveConnection(AWSKinesisSDKConnection.class).kinesisClient();
       long total = 0;
       try (CloseableIterable<PutRecordsRequestEntry> docs = CloseableIterable.ensureCloseable(requestBuilder().build(getPartitionKey(), msg))) {
         int count = 0;
@@ -138,11 +138,9 @@ public class KinesisSDKStreamProducer extends ProduceOnlyProducerImp {
     }
   }
 
-  private void doSend(AmazonKinesis kinesisClient, String endpoint, List <PutRecordsRequestEntry> putRecordsRequestEntryList){
-    PutRecordsRequest putRecordsRequest  = new PutRecordsRequest();
-    putRecordsRequest.setStreamName(endpoint);
-    putRecordsRequest.setRecords(putRecordsRequestEntryList);
-    PutRecordsResult putRecordsResult = kinesisClient.putRecords(putRecordsRequest);
+  private void doSend(KinesisClient kinesisClient, String endpoint, List <PutRecordsRequestEntry> putRecordsRequestEntryList){
+    PutRecordsRequest putRecordsRequest  = PutRecordsRequest.builder().streamName(endpoint).records(putRecordsRequestEntryList).build();
+    PutRecordsResponse putRecordsResult = kinesisClient.putRecords(putRecordsRequest);
     log.trace("PutRecordResults: {}", putRecordsResult);
   }
