+++ interlok-aws2-kinesis-sdk/src/main/java/com/adaptris/aws/kinesis/RequestBuilder.java
@@ -1,4 +1,4 @@
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.CoreException;
