+++ interlok-aws2-kinesis-sdk/src/test/java/com/adaptris/aws/kinesis/KinesisSDKStreamProducerTest.java
@@ -1,4 +1,4 @@
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
